Refactor database selection and centralize package management (#25)
* Refactor database selection and centralize package management

The code was changed to improve the way different databases are selected by changing "useMSSQL" and "usePOSTGRE" to "mssql" and "postgre", and adjusting the related conditions. Additionally, all package references for the Api project have been centralized into a new Directory.Packages.props file to simplify management and enhance consistency. A few redundant import lines in ModelConfiguration have also been removed.

* Refactor project configuration and update codebase

Moved the version management of packages to a central location and refactored the property group in various project files for better interoperability and version management. This update also includes modifications in database setup according to the project requirements. Also, several naming conventions were improved for clarity and standardization.

* Remove unnecessary white spaces in project files

The commit includes the removal of unused white spaces or empty lines in the Directory.Packages.props and Templates.csproj files. This is part of an effort to declutter and standardize the format of the project files for better readability and maintenance.

* Update packages and organize dependencies

Revised and organized the package dependencies in Database and Project directories. Unnecessary packages were removed in favor of more specific ones. The conditional statements were also added based on the database type to effectively manage the Microsoft and Npgsql packages.

* Update PostgreSQL switch in documentation

The command-line switch for using PostgreSQL in the project documentation has been updated. The previous switch `--usePOSTGRE` has been changed to `--postgre` across multiple files to ensure consistency. This is reflected in the `model-configuration.md`, `api-solution.md`, `database-project.md`, `context-factory.md`, and `context.md` files.

* Update package versions and clean up Directory.Packages.props

Removed unnecessary line breaks and replaced PackageReference tags with corresponding PackageVersion tags in Directory.Packages.props. Also, added a new Directory.Build.props file in the tests directory. This improves alignment of package versions across all project files and simplifies the packages property file.

* Update project dependencies and database configuration documentation

Removed unnecessary spaces in dependency configuration file and updated the Entity Project section in the documentation for clarity on project structure and relationships. Added a new documentation file explaining the steps to implement dependency injection in the project. Also, added these changes to the changelog.

* Update changelog with recent modifications

The changelog documentation has been updated to reflect recent changes made to the Database project. The changes regarding the implementation of Dependency Injection capability to the database projects and the updated documentation to incorporate these modifications have been included.

---------

Co-authored-by: Gary Woodfine <lnb0l9dc@duck.com>